Split projects and project component
