+++ src/Projects/projects.scss
@@ -1,0 +1,75 @@
+main {
+  display: grid;
+  grid-template-columns: 1fr;
+  grid-template-rows: 10% 1fr 2fr;
+  grid-template-areas:
+    'nav'
+    'inputs'
+    'projects';
+  min-height: 100vh;
+}
+
+h1 {
+  display: flex;
+  flex-direction: flex-start;
+  grid-area: nav;
+  margin-left: 20px;
+}
+
+.project-inputs {
+  display: flex;
+  flex-direction: column;
+  grid-area: inputs;
+  justify-content: center;
+  margin: 0px auto;
+  width: 50%;
+}
+
+.project-input {
+  font-size: 1em;
+  height: 50px;
+  margin-bottom: 5px;
+  padding-left: 10px;
+}
+
+.project-button {
+  cursor: pointer;
+  font-size: 1.2em;
+  height: 50px;
+  margin-top: 5px;
+    &:hover {
+      background-color: lightgrey;
+    }
+}
+
+.project-display {
+  align-items: center;
+  background-color: #D9DCDF;
+  display: grid;
+  grid-area: projects;
+  grid-template-columns: 1fr 1fr 1fr;
+  grid-template-rows: 1fr 1fr;
+  grid-gap: 20px;
+  padding: 20px;
+    &:hover .project-prompt {
+      color: black;
+    }
+}
+
+.project-prompt {
+  color: #D9DCDF;
+  left: 42%;
+  position: absolute;
+}
+
+.project {
+  align-items: center;
+  border: 1px solid;
+  cursor: pointer;
+  display: flex;
+  height: 75px;
+  justify-content: center;
+   &:hover {
+     background-color: #fff;
+   }
+}
